+++ Changelog.md
@@ -10,6 +10,7 @@ WhateverGreen Changelog
 - Added `applbkl` aliases to IORegistry (data, 32-bit)
 - Added `applbkl-name` and `applbkl-data` IORegistry data keys to provide custom backlight data
 - Fixed applying CoreFP patches on Apple firmware, when they are not needed
+- Added `shikigva=16` (repurposed) property to use AMD hardware DRM decoder in select apps
 
 #### v1.3.4
 - Added support for disabled AppleGraphicsDevicePolicy in AMD drivers on 10.15.1